+++ project_config/webpack/util/make_module.js
@@ -21,12 +21,18 @@ module.exports = (config) => {
     var rules = [babelLoader, htmlLoader];
 
     if (config.packCssSeparately) {
-        if (config.enableCssModules) {
+        if (config.enableCssModules && config.excludeCssNodeModules) {
             rules.push(
                 makeExtractCssLoader(!1, !0, !1),
                 makeExtractCssLoader(!0, !1, !0),
                 makeExtractLessLoader(!1, !0, !1),
                 makeExtractLessLoader(!0, !1, !0)
+            );
+        }
+        else if (config.enableCssModules) {
+            rules.push(
+                makeExtractCssLoader(!0, !1, !1),
+                makeExtractLessLoader(!0, !1, !1)
             );
         }
         else {
@@ -37,12 +43,18 @@ module.exports = (config) => {
         }
     }
     else {
-        if (config.enableCssModules) {
+        if (config.enableCssModules && config.excludeCssNodeModules) {
             rules.push(
                 makeCssLoader(!1, !0, !1),
                 makeCssLoader(!0, !1, !0),
                 makeLessLoader(!1, !0, !1),
                 makeLessLoader(!0, !1, !0)
+            );
+        }
+        else if (config.enableCssModules) {
+            rules.push(
+                makeCssLoader(!0, !1, !1),
+                makeLessLoader(!0, !1, !1)
             );
         }
         else {