+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.2.1
 
 - modify `resolveModules` sequence.
+- announcing that the css module must be defined before js module in `resolveModules, outResolveAlias`.
+- use `transform-react-jsx` by `require("babel-plugin-transform-react-jsx")`
 
 ## 0.2.0